Add env var for parametrize google root orgunit alias for sdm roles (#16)
